Migrates Tripy to use the tensorrt dialect instead of StableHLO (#607)
Migrates Tripy to use the `tensorrt` dialect instead of StableHLO. The
latter is a lower level representation that the TensorRT network API,
meaning that we first break down high level Tripy operations and then
pattern match them back up. Using the `tensorrt` dialect allows us to go
directly to TensorRT without so many intervening layers of translation,
which reduces complexity and bugs, and improves performance.

Broadly, this change does the following:
- Removes FlatIR
- Make Trace responsible for mapping to MLIR
    - Trace operations represent MLIR operations 1:1
- Makes the frontend API responsible for composing Trace operations

---------

+++ tripy/nvtripy/frontend/ops/cast.py
@@ -1,5 +1,5 @@
 #
-# SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -36,7 +36,7 @@ from nvtripy.utils import wrappers
     dtype_exceptions=[
         {"T1": "float8", "T2": "int4"},
         {"T1": "float8", "T2": "int8"},
-        {"T1": "float8", "T2": "int64"},
+        {"T1": "int8", "T2": "float8"},
         {"T1": "int4", "T2": "float8"},
         {"T1": "int4", "T2": "int8"},
         {"T1": "int4", "T2": "int64"},
